requirements.txt
absl-py==0.9.0
adjustText==0.7.3
aiohttp==3.6.2
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.12
alembic==1.4.3
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
art==5.1
asn1crypto==0.24.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==20.3.0
autograd==1.3
autograd-gamma==0.5.0
ax-platform==0.1.19
Babel==2.9.0
backcall==0.2.0
bayesian-optimization==1.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bioinfokit==2.0.8
bitstring==3.1.6
black==20.8b1
bleach==3.3.0
blessings==1.7
blist==1.3.6
bokeh==2.1.1
botorch==0.4.0
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.5
chainer==7.7.0
chainercv==0.13.1
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cliff==3.4.0
cligj==0.7.2
cloudpickle==1.5.0
cmaes==0.6.1
cmd2==1.3.10
colorama==0.4.3
colorcet==2.0.2
colorful==0.5.4
colorlog==4.2.1
colorlover==0.3.0
configparser==5.0.2
construct==2.10.54
cryptography==2.7
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.13
dareblopy==0.0.3
dask==2.22.0
datashader==0.11.0
datashape==0.5.2
deap==1.3.0
decorator==4.4.2
defusedxml==0.7.0
descartes==1.1.0
diffdist==0.1
diskcache==4.1.0
distributed==2.22.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
easyfsl==0.2.0
ecdsa==0.13.2
einops==0.3.0
entrypoints==0.3
facets-overview==1.0.0
fcn==6.4.17
filelock==3.0.12
Fiona==1.8.20
Flask==1.1.2
flatbuffers==2.0
flax==0.3.2
fonttools==4.29.0
formulaic==0.2.3
fsspec==0.8.0
future==0.17.1
gast==0.2.2
gdown==3.12.0
gdrive==0.0.8
geojson==2.5.0
geopandas==0.9.0
gin==0.1.6
gin-config==0.5.0
girder-client==3.1.0
gitdb==4.0.7
GitPython==3.1.18
google==3.0.0
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.52.0
gpustat==0.6.0
gpytorch==1.4.0
grpcio==1.27.2
h2o==3.32.1.3
h5py==2.10.0
HeapDict==1.0.1
hiredis==1.1.0
histolab==0.2.6
holoviews==1.13.3
httplib2==0.20.4
hyperopt==0.2.5
idna==2.10
image-classifiers==1.0.0
image-slicer==0.3.0
imageio==2.8.0
imagesize==1.1.0
imbalanced-learn==0.7.0
imblearn==0.0
imgaug==0.4.0
importlib-metadata==3.7.0
importlib-resources==5.1.0
imutils==0.5.3
install==1.3.1
interface-meta==1.2.3
intervaltree==3.0.2
ipaddress==1.0.22
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jax==0.2.9
jaxlib==0.1.68
jedi==0.18.0
jeepney==0.7.1
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kneed==0.7.0
liac-arff==2.4.0
lifelines==0.26.0
linformer==0.2.1
llvmlite==0.34.0
locket==0.2.0
loguru==0.5.3
lxml==4.7.1
Mako==1.1.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-venn==0.11.6
mistune==0.8.4
mizani==0.7.3
mmdnn==0.3.1
mock==3.0.5
modAL==0.4.1
more-itertools==7.2.0
mpi4py==3.0.2
mpmath==1.1.0
msgpack==1.0.0
multidict==4.7.6
multipledispatch==0.6.0
munch==2.5.0
mypy-extensions==0.4.3
natsort==7.0.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netaddr==0.7.19
netifaces==0.10.9
networkx==2.4
nibabel==3.1.1
nimfa==1.4.0
nose==1.3.7
notebook==6.4.8
numba==0.51.0
numpy==1.17.3
nvidia-dali-cuda100==0.23.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.10
opencensus-context==0.1.1
opencv-python==4.2.0.34
openslide-python==1.1.2
opt-einsum==3.2.0
optuna==2.1.0
ordered-set==4.0.2
packaging==20.9
palettable==3.3.0
pandas==0.25.3
pandocfilters==1.4.3
panel==0.9.7
param==1.9.3
parameterized==0.7.4
paramiko==2.6.0
parso==0.8.1
partd==1.1.0
pathlib2==2.3.4
pathspec==0.8.0
pathtools==0.1.2
patsy==0.5.1
paycheck==1.0.2
pbr==5.4.3
pickleshare==0.7.5
Pillow==7.2.0
pipreqs==0.4.11
plotly==4.8.1
plotnine==0.8.0
pluggy==0.13.0
pma-python==2.0.0.129
prefetch-generator==1.0.1
prettytable==0.7.2
progressbar2==3.53.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.11.3
psutil==5.6.3
ptyprocess==0.7.0
py==1.8.0
py-expression-eval==0.3.9
py-spy==0.3.3
pyaml==20.4.0
pyasn1==0.4.7
pyasn1-modules==0.2.8
pycm==3.1
pycocotools==2.0.2
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.14.1
pyct==0.4.6
Pygments==2.8.0
pykeepass==4.0.1
pymongo==3.11.2
PyNaCl==1.3.0
pynndescent==0.5.6
pyparsing==2.4.7
pyperclip==1.8.0
pyproj==3.1.0
pyrsistent==0.17.3
pyslurm==20.11.8.0
PySocks==1.7.1
pytest==5.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-utils==2.5.6
pytz==2021.1
pyvips==2.1.16
pyviz-comms==0.7.6
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
ray==0.9.0.dev0
redis==3.4.1
regex==2020.7.14
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.6
scandir==1.10.0
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-optimize==0.7.4
scikit-plot==0.3.7
scipy==1.3.1
seaborn==0.10.1
SecretStorage==3.2.0
selective-search==0.1.2.2
Send2Trash==1.8.0
sentry-sdk==1.3.0
setuptools-scm==3.3.3
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
skorch==0.8.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==1.9.1
song-vis==1.3.2.7
sortedcontainers==2.2.2
soupsieve==2.0.1
spams==2.6.2.5
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.1.2
SQLAlchemy==1.3.19
staintools==2.1.2
statsmodels==0.12.2
stevedore==3.2.2
submitit==1.3.3
subprocess32==3.5.4
tabulate==0.8.3
tblib==1.7.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorlayer==2.2.3
termcolor==1.1.0
terminado==0.9.2
test-tube==0.7.5
testpath==0.4.4
textwrap3==0.9.2
tfrecord==1.11
threadpoolctl==2.1.0
tifffile==2020.6.3
timm==0.3.2
toml==0.10.1
toolz==0.10.0
torch==1.9.0
torchaudio==0.10.2
torchlars==0.1.2
torchmeta==1.8.0
torchray==1.0.0.2
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
tune-sklearn==0.0.7
typed-ast==1.4.1
typeguard==2.11.1
typing-extensions==3.7.4.3
ujson==1.35
umap-learn==0.4.6
uritemplate==3.0.1
urllib3==1.26.3
versioneer==0.21
wandb==0.10.33
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
xarray==0.16.0
xlrd==1.2.0
yarg==0.1.9
yarl==1.5.1
zict==2.0.0
python-dateutil
pytz
